requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
future==0.18.2
gunicorn==20.0.4
idna==2.10
line-bot-sdk==1.18.0
pytz==2021.1
requests==2.25.1
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.3
